requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
python-docx==1.1.0
openai==1.6.1
PyPDF2==3.0.1
pdf2image==1.16.3
requests==2.31.0